Generate cli docs
