+++ internal/cli/docs/summaraizer.md
@@ -1,0 +1,24 @@
+## summaraizer
+
+Summarizes comments
+
+### Synopsis
+
+A tool to summarize comments from various sources using AI.
+
+### Options
+
+```
+  -h, --help   help for summaraizer
+```
+
+### SEE ALSO
+
+* [summaraizer github](summaraizer_github.md)	 - Summarizes using GitHub as source
+* [summaraizer gitlab](summaraizer_gitlab.md)	 - Summarizes using GitLab as source
+* [summaraizer mistral](summaraizer_mistral.md)	 - Summarizes using Mistral AI
+* [summaraizer ollama](summaraizer_ollama.md)	 - Summarizes using Ollama AI
+* [summaraizer openai](summaraizer_openai.md)	 - Summarizes using OpenAI AI
+* [summaraizer reddit](summaraizer_reddit.md)	 - Summarizes using Reddit as source
+* [summaraizer slack](summaraizer_slack.md)	 - Summarizes using Slack as source
+
