+++ internal/cli/docs/summaraizer_github.md
@@ -1,0 +1,24 @@
+## summaraizer github
+
+Summarizes using GitHub as source
+
+### Synopsis
+
+Summarizes using GitHub as source
+
+```
+summaraizer github [flags]
+```
+
+### Options
+
+```
+  -h, --help           help for github
+      --issue string   The GitHub issue to summarize. Use the format owner/repo/issue_number.
+      --token string   The GitHub token. Only required for private repositories.
+```
+
+### SEE ALSO
+
+* [summaraizer](summaraizer.md)	 - Summarizes comments
+
