+++ internal/cli/docs/summaraizer_gitlab.md
@@ -1,0 +1,25 @@
+## summaraizer gitlab
+
+Summarizes using GitLab as source
+
+### Synopsis
+
+Summarizes using GitLab as source
+
+```
+summaraizer gitlab [flags]
+```
+
+### Options
+
+```
+  -h, --help           help for gitlab
+      --issue string   The GitLab issue to summarize. Use the format owner/repo/issue_number.
+      --token string   The GitLab token.
+      --url string     The URL of the GitLab instance. (default "https://gitlab.com")
+```
+
+### SEE ALSO
+
+* [summaraizer](summaraizer.md)	 - Summarizes comments
+
