+++ internal/cli/docs/summaraizer_reddit.md
@@ -1,0 +1,23 @@
+## summaraizer reddit
+
+Summarizes using Reddit as source
+
+### Synopsis
+
+Summarizes using Reddit as source.
+
+```
+summaraizer reddit [flags]
+```
+
+### Options
+
+```
+  -h, --help          help for reddit
+      --post string   The Reddit post to summarize. Use the URL path. Everything after reddit.com.
+```
+
+### SEE ALSO
+
+* [summaraizer](summaraizer.md)	 - Summarizes comments
+
