+++ internal/cli/docs/summaraizer_slack.md
@@ -1,0 +1,25 @@
+## summaraizer slack
+
+Summarizes using Slack as source
+
+### Synopsis
+
+Summarizes using Slack as source
+
+```
+summaraizer slack [flags]
+```
+
+### Options
+
+```
+      --channel string   The channel ID of the Slack thread.
+  -h, --help             help for slack
+      --token string     The Slack token.
+      --ts string        The timestamp of the Slack thread.
+```
+
+### SEE ALSO
+
+* [summaraizer](summaraizer.md)	 - Summarizes comments
+
